refactor decoding of user identity into separate util class
